Refactored polling out into separate classes
